requirements.txt
Django==4.2.14
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.4.0
boto3==1.34.143
django-storages==1.14.4
django-redis-cache==3.0.1
django-utils-six==2.0
meinheld==1.0.2
whitenoise==6.7.0
Cython==3.0.10
django-rest-passwordreset==1.4.1
requests==2.32.3
pandas==2.0.3
drf-yasg==1.21.7
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
coreapi==2.3.3
PyYAML==6.0.1
fpdf2==2.7.9
regex==2024.5.15
xmltodict==0.13.0
XlsxWriter==3.2.0
pyTelegramBotAPI